edits from Kirk (#93)
* Update analysis2_manuscript.tex

Here are my edits on the revision. I added some context on the vaquita in a couple places. I also switched out the word "benchmark" in a few places to make it seem less like we're claiming we did a proper benchmarking. This is in line with the spirit of the response to the last comment from the first reviewer.

* Apply suggestions from code review

Co-authored-by: Peter Ralph <petrel.harp@gmail.com>

---------

Co-authored-by: Andrew Kern <adkern@uoregon.edu>
Co-authored-by: Peter Ralph <petrel.harp@gmail.com>